+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-alpha.xx]-2025-02-05
+
+### Added
+
+- Messages API endpoints:
+  - Get all channels
+  - Get single channel
+  - Load more messages
+  - Send message
+  - Close channel
 
 ## [1.0.0-alpha.27]-2025-02-04
 
@@ -28,7 +38,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Get product by slug
   - Get latest products
   - Get product variant lookup
-
 
 ## [1.0.0-alpha.25]-2025-02-03
 
@@ -180,4 +189,3 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security
 
 - N/A
-